Add ingredient integrated with recipe search result
Note that as far as I can tell, this functionality will only fully work
when we have a matching item in the mock database already. Otherwise,
though the item is added to the grocery list, we may encounter issues
with it down the line. Other design improvements included, and BAE
intent can now accept more than one item to add with the intent extras.